arrayMethods finished
