+++ index.html
@@ -2,9 +2,11 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
+    <link rel="stylesheet" href="./styles.css">
     <title>Tech It Easy Dashboard</title>
 </head>
 <body>
 //Push nieuwe branch
+<script type="module" src="./main.js"></script>
 </body>
 </html>
